timetrack/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
drf-spectacular>=0.26.5
pytest>=7.4.3
pytest-django>=4.7.0
sentry-sdk>=1.39.1
django-filter>=23.5
django-cors-headers>=4.6.0
